Move magic commands from kernel to Jupyter project. (#522)
* Move magic commands from kernel to Jupyter project.

* Adapt tests.

* Move GetCommsRouter as well.

* Fix magic resolver initialization in tests.

* Add missing using statement.

* Update to 185752-beta.

* Disambiguate class name.

* Update test for missing magic.

* Relax overly strict test.
