+++ src/Jupyter/Magic/LsOpenMagic.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Jupyter.Core;
 using Microsoft.Quantum.IQSharp.Jupyter;
 
-namespace Microsoft.Quantum.IQSharp.Kernel
+namespace Microsoft.Quantum.IQSharp.Jupyter
 {
     /// <summary>
     ///     A magic command that lists any open namespaces and their aliases.
