+++ src/Tests/Startup.cs
@@ -53,11 +53,15 @@ namespace Tests.IQSharp
             var serviceProvider = services.BuildServiceProvider();
             serviceProvider.GetRequiredService<ITelemetryService>();
             serviceProvider.GetRequiredService<IWorkspace>().Initialization.Wait();
+            serviceProvider.AddBuiltInMagicSymbols();
             return serviceProvider;
         }
 
-        internal static T Create<T>(string workspaceFolder) =>
-            ActivatorUtilities.CreateInstance<T>(CreateServiceProvider(workspaceFolder));
+        internal static T Create<T>(string workspaceFolder)
+        {
+            var serviceProvider = CreateServiceProvider(workspaceFolder);
+            return ActivatorUtilities.CreateInstance<T>(serviceProvider);
+        }
 
         public static void AddMocks(this IServiceCollection services)
         {
