+++ src/Tool/Telemetry.cs
@@ -17,6 +17,7 @@ using Microsoft.Quantum.QsCompiler.CompilationBuilder;
 using Microsoft.Quantum.Simulation.Simulators;
 using static Microsoft.Jupyter.Core.BaseEngine;
 using Microsoft.Quantum.IQSharp.Kernel;
+using Microsoft.Quantum.IQSharp.Jupyter;
 using Microsoft.Quantum.IQSharp.AzureClient;
 using System.Collections.Generic;
 
@@ -86,7 +87,7 @@ namespace Microsoft.Quantum.IQSharp
                 TelemetryLogger.LogEvent(evt);
             };
             this.OnServiceInitialized<IReferences>();
-            this.OnServiceInitialized<IConfigurationSource>();
+            this.OnServiceInitialized<Microsoft.Quantum.IQSharp.Jupyter.IConfigurationSource>();
             this.OnServiceInitialized<IMetadataController>();
             this.OnServiceInitialized<INugetPackages>();
             this.OnServiceInitialized<IAzureClient>();
